Implement window list plugin for Polybar
An AwesomeWM-like window list for Polybar.

There are some limitations on Polybar side, so it's not a fully
identical implementation.

For example, Polybar doesn't provide an ability to fit a set of
elements into a container of a dynamic width. Therefore, the plugin
gets the max-width as the first argument. It's up to user to figure
out the best value for the max-width.

Internally, it listens to xorg events (via `xev`), which seem to be
enough to catch the window focus changes.

Current limitations:

- The window order is not the same as in the WM